fix vcov resmean

+++ src/fastcox.cpp
@@ -1060,7 +1060,7 @@ RcppExport SEXP cumsumstratasumR(SEXP ia,SEXP istrata, SEXP instrata,SEXP iallst
 	colvec a = Rcpp::as<colvec>(ia);
 	//  mat b = Rcpp::as<mat>(ib);
 	IntegerVector intstrata(istrata);
-	int nstrata = Rcpp::as<int>(instrata);
+	unsigned nstrata = Rcpp::as<unsigned>(instrata);
 	int allstrata = Rcpp::as<int>(iallstrata);
 	unsigned n = a.n_rows;
 
@@ -1075,7 +1075,7 @@ RcppExport SEXP cumsumstratasumR(SEXP ia,SEXP istrata, SEXP instrata,SEXP iallst
 	for (unsigned i=0; i<n; i++) {
 		ss=intstrata(i);
 		// valid strata update
-		if ((first>0.1) && (i>=1)&& (ss<nstrata) && (ss>=0)) {
+		if ((first>0.1) && (i>=1)&& ((unsigned)ss<nstrata) && (ss>=0)) {
 			ressqu(i) = ressqu(i - 1) + pow(a(i), 2) + 2 * a(i) * tmpsum(ss);
 		}
 		lagressum(i)=tmpsum(ss);
